split export and tests due to size, clean comments
