+++ fidesctl/tests/core/test_export.py
@@ -2,8 +2,6 @@ from os import path
 
 import pytest
 
-import pandas as pd
-
 from fidesctl.core import export
 from fideslang.models import (
     Dataset,
@@ -94,30 +92,6 @@ def test_system_records_to_export(test_sample_system_taxonomy, test_config):
 
 
 @pytest.mark.unit
-def test_dataset_data_category_rows(test_sample_dataset_taxonomy):
-
-    dataset_name = test_sample_dataset_taxonomy[0].name
-    dataset_description = test_sample_dataset_taxonomy[0].description
-    dataset_third_country_transfers = test_sample_dataset_taxonomy[
-        0
-    ].third_country_transfers
-    dataset_fides_key = test_sample_dataset_taxonomy[0].fides_key
-    test_field = test_sample_dataset_taxonomy[0].collections[0].fields[1]
-
-    dataset_rows = export.generate_data_category_rows(
-        dataset_name,
-        dataset_description,
-        test_field.data_qualifier,
-        test_field.data_categories,
-        test_field.retention,
-        dataset_third_country_transfers,
-        dataset_fides_key,
-    )
-
-    assert len(dataset_rows) == 2
-
-
-@pytest.mark.unit
 def test_dataset_records_to_export(test_sample_dataset_taxonomy):
     """
     Asserts that unique records are returned properly (including
@@ -139,58 +113,3 @@ def test_organization_records_to_export():
         [Organization(fides_key="default_organization")]
     )
     assert len(output_list) == 5
-
-
-@pytest.mark.unit
-def test_csv_export(tmpdir):
-    """
-    Asserts that the csv is successfully created
-    """
-
-    output_list = [("column_1", "column_2"), ("foo", "bar")]
-    resource_exported = "test"
-
-    exported_filename = export.export_to_csv(
-        list_to_export=output_list,
-        resource_exported=resource_exported,
-        manifests_dir=f"{tmpdir}",
-    )
-
-    exported_file = tmpdir.join(exported_filename)
-
-    assert path.exists(exported_file)
-
-
-@pytest.mark.unit
-def test_xlsx_export(tmpdir):
-    """
-    Asserts that the xlsx template is successfully copied and appended to
-    """
-    output_columns = [
-        "dataset.name",
-        "system.name",
-        "system.administrating_department",
-        "system.privacy_declaration.data_use.name",
-        "system.joint_controller",
-        "system.privacy_declaration.data_subjects",
-        "dataset.data_categories",
-        "system.privacy_declaration.data_use.recipients",
-        "system.link_to_processor_contract",
-        "third_country_combined",
-        "system.third_country_safeguards",
-        "dataset.retention",
-        "organization.link_to_security_policy",
-    ]
-
-    organization_df = pd.DataFrame()
-    joined_system_dataset_df = pd.DataFrame(columns=output_columns)
-
-    exported_filename = export.export_datamap_to_excel(
-        organization_df=organization_df,
-        joined_system_dataset_df=joined_system_dataset_df,
-        manifests_dir=f"{tmpdir}",
-    )
-
-    exported_file = tmpdir.join(exported_filename)
-
-    assert path.exists(exported_file)
